Implement new `snap::read::FrameEncoder` type
We factor out some of the code from `snap::write::FrameEncoder`, but it
might be possible to factor out some more code from
`snap::read::FrameDecoder` as well. We provide test cases to make sure
that we encode all the same files the same way that the original code
does.

We have some code which attempts to use either an internal compression
buffer or the one provided by our caller if it's big enough. We probably
want one or two more tests to make sure that this works correctly both
ways.