readme: re-run benchmarks
It looks like the benchmarks results remain relatively the same.
Although messages on the snappy mailing list suggest there may have been
a regression in the reference implementation. I haven't yet had a chance
to review any new optimizations in the reference implementation since I
ported it to Rust, so we'll have to do without any analysis for now.

Closes #6