test: tweak snappy-cpp tests
Originally, when initially building `snap`, I wrote tests that ensured
this crate was byte-for-byte identical with its compression when
compared to `snappy-cpp`. This was obviously brittle and very ambitious
of me. The idea was I wanted to make sure I was doing at least as good
as `snappy-cpp`.

Alas, time has caught up with me. I'm not sure if I have any plans to
dig back into this crate and catch back up with `snappy-cpp`. So for
now, I've tweaked the tests to just ensure that `snap` can decompress
data compressed by `snappy-cpp` and vice versa.